+++ app/Bibliography.php
@@ -2,8 +2,8 @@
 
 namespace App;
 
+use App\File\Directory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use Spatie\Activitylog\Traits\LogsActivity;
 use Spatie\Activitylog\LogOptions;
@@ -15,10 +15,6 @@ class Bibliography extends Model implements Searchable
     use LogsActivity;
 
     protected $table = 'bibliography';
-
-    protected $appends = [
-        'file_url',
-    ];
 
     /**
      * The attributes that are assignable.
@@ -159,6 +155,7 @@ class Bibliography extends Model implements Searchable
                 'abstract',
                 'address',
                 'author',
+                'doi',
                 'edition',
                 'editor',
                 'email',
@@ -187,6 +184,7 @@ class Bibliography extends Model implements Searchable
                 'author',
                 'booktitle',
                 'chapter',
+                'doi',
                 'edition',
                 'editor',
                 'email',
@@ -216,6 +214,7 @@ class Bibliography extends Model implements Searchable
             "fields" => [
                 'abstract',
                 'author',
+                'doi',
                 'email',
                 'howpublished',
                 'language',
@@ -231,6 +230,7 @@ class Bibliography extends Model implements Searchable
                 'abstract',
                 'address',
                 'author',
+                'doi',
                 'email',
                 'howpublished',
                 'language',
@@ -250,6 +250,7 @@ class Bibliography extends Model implements Searchable
                 'address',
                 'author',
                 'booktitle',
+                'doi',
                 'editor',
                 'email',
                 'language',
@@ -278,6 +279,7 @@ class Bibliography extends Model implements Searchable
                 'address',
                 'author',
                 'chapter',
+                'doi',
                 'edition',
                 'editor',
                 'email',
@@ -311,6 +313,7 @@ class Bibliography extends Model implements Searchable
                 'address',
                 'author',
                 'booktitle',
+                'doi',
                 'editor',
                 'email',
                 'language',
@@ -338,6 +341,7 @@ class Bibliography extends Model implements Searchable
                 'abstract',
                 'address',
                 'author',
+                'doi',
                 'edition',
                 'email',
                 'language',
@@ -357,6 +361,7 @@ class Bibliography extends Model implements Searchable
                 'abstract',
                 'address',
                 'author',
+                'doi',
                 'email',
                 'language',
                 'month',
@@ -379,6 +384,7 @@ class Bibliography extends Model implements Searchable
                 'abstract',
                 'address',
                 'author',
+                'doi',
                 'email',
                 'language',
                 'month',
@@ -400,6 +406,7 @@ class Bibliography extends Model implements Searchable
             "fields" => [
                 'abstract',
                 'address',
+                'doi',
                 'editor',
                 'email',
                 'language',
@@ -424,6 +431,7 @@ class Bibliography extends Model implements Searchable
                 'abstract',
                 'address',
                 'author',
+                'doi',
                 'email',
                 'institution',
                 'language',
@@ -431,7 +439,7 @@ class Bibliography extends Model implements Searchable
                 'note',
                 'number',
                 'title',
-                'type'   ,
+                'type',
                 'url',
                 'year',
             ],
@@ -530,6 +538,21 @@ class Bibliography extends Model implements Searchable
         return $strippedFields;
     }
 
+    public function unsetDisallowed(): void {
+        $typeFields = self::bibtexTypes[$this->entry_type]['fields'];
+
+        $allFields = $this->fillable;
+        foreach($allFields as $field) {
+            if(in_array($field, $typeFields)) continue;
+            if($field == 'entry_type') continue;
+            if($field == 'file') continue;
+            if($field == 'citekey') continue;
+            if($field == 'user_id') continue;
+
+            $this->{$field} = null;
+        }
+    }
+
 
     private function getEmptyFields(string $type) : array {
         $typeFields = self::bibtexTypes[$type];
@@ -551,13 +574,11 @@ class Bibliography extends Model implements Searchable
 
         $type = $fields['entry_type'];
         $filteredFields = self::stripDisallowed($fields, $type);
-        $nullMap = array_map(fn() => null, self::patchRules );
-        $nullMap = array_filter($nullMap, fn($key) => !in_array($key, ["file"]), ARRAY_FILTER_USE_KEY);
-        $allFieldsFixed = array_merge($nullMap, $filteredFields);
-        foreach($allFieldsFixed as $key => $value) {
+        foreach($filteredFields as $key => $value) {
             if(!in_array($key, $this->fillable)) throw new \Exception("Field $key is not allowed for this type of entry");
             $this->{$key} = $value;
         }
+        $this->unsetDisallowed();
 
         // updating an item does not have to update all fields
         // thus we first set all allowed keys from request and then
@@ -646,22 +667,21 @@ class Bibliography extends Model implements Searchable
         return $key;
     }
 
-    public function uploadFile($file) {
-        $this->deleteFile(true);
-
-        $filename = $this->id . "_" . $file->getClientOriginalName();
-        return $file->storeAs(
-            'bibliography',
-            $filename
-        );
+    public function deleteFileFromStorage(): bool {
+        return self::getDirectory()->delete($this->file);
     }
 
-    public function deleteFile(bool $fromStorageOnly = false) {
-        if(isset($this->file) && Storage::exists($this->file)) {
-            Storage::delete($this->file);
-        }
+    public function uploadFile($file): void {
+        $this->deleteFileFromStorage();
+        $filename = $this->id . "_" . $file->getClientOriginalName();
+        $storagePath = self::getDirectory()->store($filename, $file);
+        $this->file = $storagePath;
+        $this->save();
+    }
 
-        if(!$fromStorageOnly) {
+    public function deleteFile(): void {
+        $success = $this->deleteFileFromStorage();
+        if($success) {
             $this->file = null;
             $this->save();
         }
@@ -675,7 +695,7 @@ class Bibliography extends Model implements Searchable
         return $this->belongsToMany('App\Entity', 'references')->withPivot('description', 'attribute_id');
     }
 
-    public function getFileUrlAttribute() {
-        return isset($this->file) ? sp_get_public_url($this->file) : null;
+    public static function getDirectory() : Directory{
+        return new Directory('bibliography');
     }
 }