+++ app/Http/Controllers/BibliographyController.php
@@ -1,18 +1,19 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Bibliography;
 use Illuminate\Http\Request;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use RenanBr\BibTexParser\Listener;
 use RenanBr\BibTexParser\Exception\ParserException;
 use RenanBr\BibTexParser\Parser;
 use RenanBr\BibTexParser\Processor\TagNameCaseProcessor;
+use Exception;
 
 class BibliographyController extends Controller
 {
-
     // GET
     public function getBibliographyItem(int $id) {
         $user = auth()->user();
@@ -65,6 +66,18 @@ class BibliographyController extends Controller
         return response()->json($count);
     }
 
+    public function downloadFile(Request $request) {
+        $user = auth()->user();
+        if(!$user->can('bibliography_read')) {
+            return response()->json([
+                'error' => __('You do not have the permission to view attached files')
+            ], 403);
+        }
+
+        $filepath = $request->query('path');
+        return Bibliography::getDirectory()->download($filepath);
+    }
+
     // POST
 
     public function addItem(Request $request) {
@@ -76,12 +89,10 @@ class BibliographyController extends Controller
         }
 
         $this->validate($request, [
-            'entry_type' => 'required|alpha',
-            'title' => 'required|string',
+            'entry_type' => 'required|alpha|bibtex_type',
             'file' => 'file',
         ]);
 
-        $file = $request->file('file');
         $bib = new Bibliography();
         $success = $bib->fieldsFromRequest($request->except('file'), $user);
         if(!$success) {
@@ -90,14 +101,21 @@ class BibliographyController extends Controller
             ], 422);
         }
 
-        if(isset($file)) {
-            $bib->file = $bib->uploadFile($file);
-            $bib->user_id = $user->id;
-            $bib->save();
+        if($request->has('file')) {
+            $file = $request->file('file');
+            $bib->uploadFile($file);
         }
-        $bib = Bibliography::find($bib->id);
 
-        return response()->json($bib, 201);
+        // Does this make sense? It seems none of the methods throws an exception [VR]
+        // try {
+        //     if($file) $bib->uploadFile($file);
+        // } catch(Exception $e) {
+        //     info($e->getMessage());
+        //     return response()->json([
+        //         'error' => $e->getMessage()
+        //     ], 400);
+        // }
+        return response()->json($bib->refresh(), 201);
     }
 
     public function importBibtex(Request $request) {
@@ -110,7 +128,6 @@ class BibliographyController extends Controller
         $this->validate($request, [
             'file' => 'required|file'
         ]);
-
 
         // TODO:: This is too much logic for a controller. This should be moved inside the Model!
 
@@ -289,24 +306,21 @@ class BibliographyController extends Controller
                 'error' => __('This bibliography item does not exist')
             ], 400);
         }
-        $file = $request->file('file');
-        $success = $bib->fieldsFromRequest($request->except(['file', 'delete_file']), $user);
 
+        $success = $bib->fieldsFromRequest($request->except(['file', 'delete_file']), $user);
         if(!$success) {
             return response()->json([
                 'error' => __('At least one required field is not set')
             ], 422);
         }
+
+        $file = $request->file('file');
         if(isset($file)) {
             $bib->file = $bib->uploadFile($file);
-            $bib->user_id = $user->id;
-            $bib->save();
         } else if($request->has('delete_file') && sp_parse_boolean($request->get('delete_file'))) {
             $bib->deleteFile();
         }
-        $bib = Bibliography::find($bib->id);
-
-        return response()->json($bib);
+        return response()->json($bib->refresh());
     }
 
     // DELETE
@@ -327,7 +341,7 @@ class BibliographyController extends Controller
         }
 
         // Only delete file from storage to not fire 'save' event on item
-        $bib->deleteFile(true);
+        $bib->deleteFileFromStorage();
         $bib->delete();
 
         return response()->json(null, 204);