+++ app/Http/Controllers/UserController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Attribute;
 use App\Entity;
-use App\File;
 use App\Permission;
 use App\Role;
 use App\User;
@@ -12,16 +11,16 @@ use App\Http\Controllers\Controller;
 use App\Plugin;
 use App\RolePreset;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Sleep;
 use Illuminate\Validation\ValidationException;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
-class UserController extends Controller
-{
+class UserController extends Controller {
     public function __construct() {
         $this->middleware('auth:sanctum', ['except' => ['login']]);
     }
@@ -149,6 +148,11 @@ class UserController extends Controller
         return response()->json($groups);
     }
 
+    public function downloadAvatar(Request $request): Response|BinaryFileResponse {
+        $filepath = $request->query('path');
+        return User::getDirectory()->download($filepath);
+    }
+
     // POST
 
     public function login(Request $request) {
@@ -239,10 +243,7 @@ class UserController extends Controller
         }
 
         $file = $request->file('file');
-        $path = $user->uploadAvatar($file);
-        $user->avatar = $path;
-        $user->save();
-
+        $user->uploadAvatar($file);
         // return user without roles relation
         $user->unsetRelation('roles');
 
@@ -558,11 +559,7 @@ class UserController extends Controller
             ], 400);
         }
 
-        if(isset($user->avatar)) {
-            Storage::delete($user->avatar);
-        }
-        $user->avatar = null;
-        $user->save();
+        $user->deleteAvatar();
         return response()->json(null, 204);
     }