+++ routes/web.php
@@ -17,6 +17,9 @@ use Illuminate\Support\Facades\Auth;
 Route::get('/welcome', 'HomeController@welcome')->name('welcome');
 Route::get('/open', 'HomeController@external')->name('external');
 
+Route::middleware('auth:sanctum')->get('/download/avatar', 'UserController@downloadAvatar');
+Route::middleware('auth:sanctum')->get('/download/bibliography', 'BibliographyController@downloadFile');
+
 Auth::routes(["middleware" => ["auth:sanctum"]]);
 
 Route::get('/', 'HomeController@index')->name('home');